+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - added `folio:content_templates:idp_migrate_to_per_site` rake task to migrate existing content templates to per-site
 - added `folio:content_templates:remove_siteless` rake task to remove siteless content templates
 - added `upload` icon
+- added `active` column to `Folio::EmailTemplate` and possibility to disable specific email templates
 
 ## [6.2.5] - 2025-03-26